started stubs for path ordering
need some way of picking a good path from starting point to ending point
and right now my plan is to count the number of rare fruits on the path
and rate those as better than ones with less rare fruits. rare fruit is
defined as a fruit with a low win count.